requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# ML & Data Science
scikit-learn==1.3.2
numpy==1.26.3
scipy==1.11.4
sentence-transformers==2.2.2
rouge-score==0.1.2
transformers==4.35.2
torch==2.1.1

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
python-json-logger==2.0.7

# API Documentation
drf-spectacular==0.26.5

# Pagination & Validation
djangorestframework-pagination==1.0.7
marshmallow==3.20.1
